refactor protocol messages by domain
